All ternaries refactores with components
